`fitting.py` and `models.py` have been moved into a new `modeling` folder
