+++ petrofit/modeling/__init__.py
@@ -1,0 +1,2 @@
+from .fitting import *
+from .models import *
